# backend/src/syntara/api/execution_plane_facade.py
"""Execution Plane API endpoints."""

from __future__ import annotations

from datetime import datetime  # noqa: TC003 — Pydantic resolves model annotations at runtime
from typing import Annotated, Literal
from uuid import UUID  # noqa: TC003 — Pydantic resolves model annotations at runtime

import structlog
from fastapi import Depends, HTTPException, Query, Request, status
from pydantic import BaseModel, Field
from sqlmodel import col, select
from sqlmodel.ext.asyncio.session import AsyncSession  # noqa: TC002 — FastAPI resolves dependency annotations
from temporalio.exceptions import ApplicationError

from syntara.authz.dependencies import ProjectScopeFilter
from syntara.authz.engine import AllowedProjectsResult  # noqa: TC001 — FastAPI resolves dependency annotations
from syntara.core.config.base import get_settings
from syntara.core.database.session import get_db
from syntara.core.syntara_router import NO_PERMISSION, SyntaraRouter
from syntara.execution_plane.client import ExecutionPlaneHttpClient, ExecutionPlaneUnavailableError
from syntara.workflows.executions_router import get_temporal_execution_service
from syntara.workflows.models.activity_execution import ActivityExecution
from syntara.workflows.models.execution import Execution
from syntara.workflows.workflow_engine.services.temporal_execution_service import (
    TemporalExecutionService,  # noqa: TC001 — FastAPI resolves dependency annotations
)

logger = structlog.stdlib.get_logger(__name__)

router = SyntaraRouter(prefix="/api/execution_plane/v1", tags=["Execution Plane"])

_target_scope = ProjectScopeFilter("execution_target", "read")
_work_item_scope = ProjectScopeFilter("work_item", "read")


class ExecutionTargetListResponse(BaseModel):
    """Paginated list response for ExecutionTarget."""

    resources: list[ExecutionTargetFacadeRead]
    next: str | None = None
    prev: str | None = None
    total: int | None = None


class WorkItemListResponse(BaseModel):
    """Paginated list response for WorkItem."""

    resources: list[WorkItemFacadeRead]
    next: str | None = None
    prev: str | None = None
    total: int | None = None


class KubernetesPlacement(BaseModel):
    """Kubernetes scheduling metadata returned by the Execution Plane."""

    type: Literal["kubernetes"]
    namespace: str = Field(min_length=1, max_length=63, pattern=r"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$")
    node_selectors: list[str] = Field(default_factory=list)
    tolerations: list[str] = Field(default_factory=list)


class RHELPlacement(BaseModel):
    """RHEL placement marker returned by the Execution Plane."""

    type: Literal["rhel"]


class ExecutionTargetFacadeRead(BaseModel):
    """Safe target fields available to AO's authorized management API."""

    id: UUID
    cluster_id: UUID
    name: str
    backend_type: str
    endpoint: str
    placement: Annotated[KubernetesPlacement | RHELPlacement, Field(discriminator="type")]
    status: str
    enabled: bool
    is_default: bool
    status_message: str | None = None
    labels: dict[str, str]
    created_at: datetime
    last_ran_at: datetime | None = None


class WorkItemFacadeRead(BaseModel):
    """Safe work state; internal tokens and storage metadata stay private."""

    id: UUID
    status: str
    result: dict[str, object] | None = None
    created_at: datetime
    claimed_at: datetime | None = None
    completed_at: datetime | None = None


class EPCompletionEvent(BaseModel):
    """Authenticated result event delivered by the independent EP service."""

    event_id: UUID
    event_schema_version: Literal[1]
    client_id: str
    work_id: UUID
    state_revision: int
    status: Literal["completed", "failed", "cancelled", "reconciliation_required"]
    result: dict[str, object]
    completed_at: datetime


@router.get(
    "/execution_targets",
    operation_id="list_execution_targets",
    summary="List execution targets",
    description="Retrieve registered execution targets.",
)
async def list_execution_targets(
    allowed_projects: Annotated[AllowedProjectsResult, Depends(_target_scope)],
    limit: int = Query(default=20, ge=1, le=100),
) -> ExecutionTargetListResponse:
    """Read registered execution targets from EP after AO permission checks."""
    try:
        async with ExecutionPlaneHttpClient() as client:
            items = await _list_ep_resources(
                client,
                operation="execution_targets",
                allowed_projects=allowed_projects,
                limit=limit,
            )
    except ExecutionPlaneUnavailableError as exc:
        raise HTTPException(
            status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
            detail="Execution Plane unavailable",
        ) from exc
    logger.info("Listed execution targets", count=len(items))
    return ExecutionTargetListResponse(resources=[ExecutionTargetFacadeRead.model_validate(item) for item in items])


@router.get(
    "/work_items",
    operation_id="list_work_items",
    summary="List work items",
    description="Retrieve work items.",
)
async def list_work_items(
    allowed_projects: Annotated[AllowedProjectsResult, Depends(_work_item_scope)],
    limit: int = Query(default=20, ge=1, le=100),
) -> WorkItemListResponse:
    """Read work-item state from EP after AO permission checks."""
    try:
        async with ExecutionPlaneHttpClient() as client:
            items = await _list_ep_resources(
                client,
                operation="work_items",
                allowed_projects=allowed_projects,
                limit=limit,
            )
    except ExecutionPlaneUnavailableError as exc:
        raise HTTPException(
            status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
            detail="Execution Plane unavailable",
        ) from exc
    logger.info("Listed work items", count=len(items))
    return WorkItemListResponse(resources=[WorkItemFacadeRead.model_validate(item) for item in items])


async def _list_ep_resources(
    client: ExecutionPlaneHttpClient,
    *,
    operation: Literal["execution_targets", "work_items"],
    allowed_projects: AllowedProjectsResult,
    limit: int,
) -> list[dict[str, object]]:
    """List EP resources after AO permission checks. EP has no project scope."""
    if not allowed_projects.all_projects and not allowed_projects.project_ids:
        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="No project-scoped access to EP resources")
    if operation == "work_items":
        return await client.list_work_items(limit=limit)
    return await client.list_execution_targets(limit=limit)


def _ep_failure_error(status_value: str, result: dict[str, object]) -> ApplicationError:
    """Translate a non-completed EP status into a non-retryable activity failure."""
    if status_value == "cancelled":
        return ApplicationError(
            "Execution Plane work was cancelled before execution",
            result,
            type="ExecutionPlaneWorkCancelled",
            non_retryable=True,
        )
    if status_value == "reconciliation_required":
        return ApplicationError(
            "Execution Plane could not determine the work outcome",
            result,
            type="ExecutionPlaneReconciliationRequired",
            non_retryable=True,
        )
    error_message = str(result.get("error", "Execution Plane work failed"))
    error_type = str(result.get("error_type", "ScriptExecutionError"))
    return ApplicationError(error_message, result, type=error_type, non_retryable=True)


@router.post(
    "/events",
    status_code=status.HTTP_202_ACCEPTED,
    dependencies=[NO_PERMISSION],
    operation_id="accept_execution_plane_event",
    summary="Accept an Execution Plane completion event",
)
async def accept_execution_plane_event(
    request: Request,
    event: EPCompletionEvent,
    session: Annotated[AsyncSession, Depends(get_db)],
    temporal_service: Annotated[TemporalExecutionService | None, Depends(get_temporal_execution_service)],
) -> dict[str, str]:
    """Complete the AO async activity named by an EP callback; EP retries until this returns 2xx."""
    # The work item UUID is ActivityExecution.id, so AO resolves the activity's
    # (temporal_workflow_id, temporal_activity_id) and completes it the same way approval and
    # agentic callbacks do — no task token or AO-side dispatch state. Re-delivery is idempotent:
    # an already-resolved activity is a Temporal no-op.
    settings = get_settings()
    if (
        not getattr(request.state, "is_cert_authenticated", False)
        or getattr(request.state, "cert_cn", None) != settings.ep_callback_service_cn
    ):
        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Execution Plane service identity required")
    if temporal_service is None:
        raise HTTPException(status_code=status.HTTP_503_SERVICE_UNAVAILABLE, detail="Temporal is unavailable")

    row = (
        await session.exec(
            select(Execution.temporal_workflow_id, ActivityExecution.temporal_activity_id)
            .join(Execution, col(ActivityExecution.execution_id) == col(Execution.id))
            .where(col(ActivityExecution.id) == event.work_id)
        )
    ).first()
    if row is None:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No matching AO activity for this work item")
    temporal_workflow_id, temporal_activity_id = row

    if event.status == "completed":
        await temporal_service.complete_async_activity(
            temporal_workflow_id=temporal_workflow_id,
            activity_id=temporal_activity_id,
            result=event.result,
        )
    else:
        await temporal_service.fail_async_activity(
            temporal_workflow_id=temporal_workflow_id,
            activity_id=temporal_activity_id,
            error=_ep_failure_error(event.status, event.result),
        )
    logger.info("Delivered Execution Plane completion event", event_id=str(event.event_id), status=event.status)
    return {"status": "accepted"}