# backend/src/syntara/schemas/execution_plane/openapi.yaml
---
openapi: 3.1.0
info:
  title: Execution Plane API
  description: Execution Plane operator endpoints — observe execution targets and work items.
  version: 0.1.0
paths:
  /api/execution_plane/v1/execution_targets:
    get:
      tags:
        - Execution Plane
      summary: List execution targets
      description: Retrieve registered execution targets.
      operationId: list_execution_targets
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionTargetListResponse'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-app-permission:
        resource: execution_target
        action: read
  /api/execution_plane/v1/work_items:
    get:
      tags:
        - Execution Plane
      summary: List work items
      description: Retrieve work items.
      operationId: list_work_items
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemListResponse'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-app-permission:
        resource: work_item
        action: read
  /api/execution_plane/v1/events:
    post:
      tags:
        - Execution Plane
      summary: Accept an Execution Plane completion event
      description: Complete the AO async activity named by an EP callback; EP retries until this returns 2xx.
      operationId: accept_execution_plane_event
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EPCompletionEvent'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Accept Execution Plane Event
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-app-permission:
        resource: null
        action: null
components:
  schemas:
    EPCompletionEvent:
      properties:
        event_id:
          type: string
          format: uuid
          title: Event Id
        event_schema_version:
          type: integer
          const: 1
          title: Event Schema Version
        client_id:
          type: string
          title: Client Id
        work_id:
          type: string
          format: uuid
          title: Work Id
        state_revision:
          type: integer
          title: State Revision
        status:
          type: string
          enum:
            - completed
            - failed
            - cancelled
            - reconciliation_required
          title: Status
        result:
          additionalProperties: true
          type: object
          title: Result
        completed_at:
          type: string
          format: date-time
          title: Completed At
      type: object
      required:
        - event_id
        - event_schema_version
        - client_id
        - work_id
        - state_revision
        - status
        - result
        - completed_at
      title: EPCompletionEvent
      description: Authenticated result event delivered by the independent EP service.
    ExecutionTargetFacadeRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        cluster_id:
          type: string
          format: uuid
          title: Cluster Id
        name:
          type: string
          title: Name
        backend_type:
          type: string
          title: Backend Type
        endpoint:
          type: string
          title: Endpoint
        placement:
          oneOf:
            - $ref: '#/components/schemas/KubernetesPlacement'
            - $ref: '#/components/schemas/RHELPlacement'
          title: Placement
          discriminator:
            propertyName: type
            mapping:
              kubernetes: '#/components/schemas/KubernetesPlacement'
              rhel: '#/components/schemas/RHELPlacement'
        status:
          type: string
          title: Status
        enabled:
          type: boolean
          title: Enabled
        is_default:
          type: boolean
          title: Is Default
        status_message:
          anyOf:
            - type: string
            - type: 'null'
          title: Status Message
        labels:
          additionalProperties:
            type: string
          type: object
          title: Labels
        created_at:
          type: string
          format: date-time
          title: Created At
        last_ran_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Last Ran At
      type: object
      required:
        - id
        - cluster_id
        - name
        - backend_type
        - endpoint
        - placement
        - status
        - enabled
        - is_default
        - labels
        - created_at
      title: ExecutionTargetFacadeRead
      description: Safe target fields available to AO's authorized management API.
    ExecutionTargetListResponse:
      properties:
        resources:
          items:
            $ref: '#/components/schemas/ExecutionTargetFacadeRead'
          type: array
          title: Resources
        next:
          anyOf:
            - type: string
            - type: 'null'
          title: Next
        prev:
          anyOf:
            - type: string
            - type: 'null'
          title: Prev
        total:
          anyOf:
            - type: integer
            - type: 'null'
          title: Total
      type: object
      required:
        - resources
      title: ExecutionTargetListResponse
      description: Paginated list response for ExecutionTarget.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    KubernetesPlacement:
      properties:
        type:
          type: string
          const: kubernetes
          title: Type
        namespace:
          type: string
          maxLength: 63
          minLength: 1
          pattern: ^[a-z0-9]([-a-z0-9]*[a-z0-9])?$
          title: Namespace
        node_selectors:
          items:
            type: string
          type: array
          title: Node Selectors
        tolerations:
          items:
            type: string
          type: array
          title: Tolerations
      type: object
      required:
        - type
        - namespace
      title: KubernetesPlacement
      description: Kubernetes scheduling metadata returned by the Execution Plane.
    RHELPlacement:
      properties:
        type:
          type: string
          const: rhel
          title: Type
      type: object
      required:
        - type
      title: RHELPlacement
      description: RHEL placement marker returned by the Execution Plane.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    WorkItemFacadeRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        status:
          type: string
          title: Status
        result:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Result
        created_at:
          type: string
          format: date-time
          title: Created At
        claimed_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Claimed At
        completed_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Completed At
      type: object
      required:
        - id
        - status
        - created_at
      title: WorkItemFacadeRead
      description: Safe work state; internal tokens and storage metadata stay private.
    WorkItemListResponse:
      properties:
        resources:
          items:
            $ref: '#/components/schemas/WorkItemFacadeRead'
          type: array
          title: Resources
        next:
          anyOf:
            - type: string
            - type: 'null'
          title: Next
        prev:
          anyOf:
            - type: string
            - type: 'null'
          title: Prev
        total:
          anyOf:
            - type: integer
            - type: 'null'
          title: Total
      type: object
      required:
        - resources
      title: WorkItemListResponse
      description: Paginated list response for WorkItem.
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT token authentication
      scheme: bearer
      bearerFormat: JWT